microservice/requirements.txt
alembic
gunicorn
uvicorn
python-dotenv
fastapi==0.68.1
sqlalchemy==1.4.25
psycopg2-binary==2.8.6
requests